+++ Payments/src/main/java/io/github/LuizYokoyama/Payments/repository/EntryRepository.java
@@ -9,6 +9,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.UUID;
@@ -27,6 +28,6 @@ public interface EntryRepository extends JpaRepository<EntryEntity, UUID> {
             WHERE e.entryStatus = io.github.LuizYokoyama.Payments.entity.EntryStatus.DONE 
             AND e.entryDateTime >= a.aggregationDateTime AND a.accountId = :accountId
             """ )
-    Float aggregateBalanceSince(@Param("accountId") int accountId);
+    BigDecimal aggregateBalanceSince(@Param("accountId") int accountId);
 
 }
